+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-02-19 16:25:26
+
+### Changed Files
+- `src/features/recruiting/components/RecruitBottomPanel.tsx`
+
+### Statistics
+```
+ 1 file changed, 207 insertions(+), 55 deletions(-)
+```
+
 ## [Checkpoint] - 2026-02-19 14:50:53
 
 ### Changed Files